requirements.txt
matplotlib~=3.8.3
networkx~=3.2.1
numpy~=1.26.4
scikit-learn~=1.4.2
pandas~=2.2.1
pillow~=10.2.0
keyboard~=0.13.5
pygame~=2.5.2
pyside6~=6.7.0
setuptools~=68.2.2
cython~=3.0.10